% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/boot.CI.R
\name{boot.CI}
\alias{boot.CI}
\title{Bootstrap confidence intervals of mean}
\usage{
boot.CI(x, alpha = 0.05, CI.type = "all")
}
\arguments{
\item{x}{a vector of observation}

\item{alpha}{significance level (default: 0.05)}

\item{CI.type}{type of CI required (default: "all")}
}
\value{
\code{boot.CI} return list of confidence intervals of mean (\code{CI.percent}: percentile, \code{CI.BC}:bias-corrected and \code{CI.BCa}: bias-corrected and accelerated)
}
\description{
Bootstrap confidence intervals of mean
}
\examples{

set.seed(12)
boot.CI(rnorm(1000, mean=0, sd=1), alpha=0.05, CI.type="per") # example of wrong input for type
boot.CI(rnorm(1000, mean=0, sd=1), alpha=0.05, CI.type="all") # require all type

}
