% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/f.linear.plateau.R
\name{f.linear.plateau}
\alias{f.linear.plateau}
\title{Fitting linear plateau model using multiple initial values}
\usage{
f.linear.plateau(
  d,
  start = list(a = 1, b = 1, c = 1),
  plus_minus = 100,
  n.start = 1000,
  msg = FALSE
)
}
\arguments{
\item{d}{data frame with two columns (\code{x} and \code{y})}

\item{start}{initial estimate for non-linear least square (default value: \code{ list(a = 1, b = 1, c = 1)})}

\item{plus_minus}{radius of random initial values (default: \code{100})}

\item{n.start}{total number of initial points considered (default: \code{1000})}

\item{msg}{logical flag whether printing progress}
}
\value{
\code{f.linear.plateau} returns a list of two components (if converged): \code{nls.summary}: summary of the fitted model; \code{nls.model}: nls object
}
\description{
\code{f.linear.plateau} fits linear plateau model using multiple initial values. The multiple initial values are randomly sampled in  a "cube" of parameter space.
    More precisely, linear plateau model assumes y ~ a + b * (x - c) * (x <= c).
}
\examples{

set.seed(4)
x <- rep(1:300, each=4)
a <- 8; b <- 0.05; c <- 100
y <- a + b * (x - c) * (x <= c) +
    rnorm(length(x), sd=0.1)
d <- cbind(x,y)

# a converged example:
ans <- f.linear.plateau(d, start=list(a = 7, b = 0.1, c = 150),
    plus_minus=10, n.start=10, msg=FALSE)

summary(ans$nls.model)



}
